requirements.txt
flask
boto3
langchain
langchain_core
python-dotenv
requests
beautifulsoup4
